+++ lib/lite/memoize.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-%w[version shared instance klass mixin].each do |name|
+%w[version mixin instance klass].each do |name|
   require "lite/memoize/#{name}"
 end