+++ lib/lite/memoize/mixin.rb
@@ -4,7 +4,19 @@ module Lite
   module Memoize
     module Mixin
 
-      include Lite::Memoize::Shared
+      def store
+        @store ||= {}
+      end
+
+      def caller_key(block, as: nil)
+        name = as ? [as] : block.source_location
+        return name.concat(block) if block.is_a?(Array)
+
+        block.binding.local_variables.each_with_object(name) do |local_name, array|
+          array << local_name
+          array << block.binding.local_variable_get(local_name)
+        end
+      end
 
       def memoize(method_name, args: nil, refresh: false)
         key = "#{method_name}:#{args}"