+++ spec/lite/memoize/mixin_spec.rb
@@ -17,6 +17,22 @@ RSpec.describe Lite::Memoize::Mixin do
   let(:klass) { KlassFooService.new }
   let(:foo) { klass }
 
+  describe '.store' do
+    it 'to be {}' do
+      expect(foo.store).to eq({})
+    end
+  end
+
+  describe '.caller_key' do
+    it 'to be ["random", 1, {}, []]' do
+      expect(klass.caller_key([1, {}, []], as: 'random')).to eq(['random', 1, {}, []])
+    end
+
+    it 'to be "random"' do
+      expect(klass.caller_key([], as: 'random')).to eq(['random'])
+    end
+  end
+
   describe '.memoize' do
     it 'to be same string twice' do
       old_random_string = klass.random